create a check-in time log table

+++ databases/daily_todo/daily_todo.rb
@@ -10,8 +10,8 @@ require 'faker'
 db = SQLite3::Database.new("todo.db")
 db.results_as_hash = true
 
-# create SQL command
-create_table_cmd = <<-SQL
+# create SQL command for todo table
+create_todo_table_cmd = <<-SQL
 	CREATE TABLE IF NOT EXISTS todo(
 		id INTEGER PRIMARY KEY,
 		activity VARCHAR(255),
@@ -19,8 +19,18 @@ create_table_cmd = <<-SQL
 	) 
 SQL
 
-# create todo list table
-db.execute(create_table_cmd)
+# create SQL command for time log table
+create_log_table_cmd = <<-SQL
+	CREATE TABLE IF NOT EXISTS log(
+		id INTEGER PRIMARY KEY,
+		day ARRAY,
+		time ARRAY
+	) 
+SQL
+
+# create todo list tables
+db.execute(create_todo_table_cmd)
+db.execute(create_log_table_cmd)
 
 # create method to add activities
 def add_activity(db, activity)
@@ -31,19 +41,17 @@ end
 # create method to print unfinished activities
 def print_todo_list(db)
 	todo_list = db.execute("SELECT * FROM todo WHERE done='false'")
-	puts "Here's your daily To-Do List:"
+	puts "\nHere's your daily To-Do List:"
 	puts '*~'*20+'*'
 	todo_list.each do |activity|
 		puts activity['activity']
 	end
 end
 
-# create method to wipe table clean
+# create method to wipe table clean (always double checks!)
 def delete_table(db)
 	puts ('Are you sure you want to wipe the table? (y/n)')
-	if gets.chomp == 'y'
-		db.execute("DELETE FROM todo")
-	end
+	db.execute("DELETE FROM todo") if gets.chomp == 'y'
 end
